backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
sentence-transformers==2.3.1
scikit-learn==1.4.0
redis==5.0.1
rq==1.16.1
pydantic==2.5.3
